# backend/static/openapi.yaml
openapi: "3.0.3"
info:
  title: Auth-as-a-Service API
  description: |
    Context-aware multi-factor authentication service.
    External applications authenticate via API keys and receive JWT tokens.

    ## Authentication
    - **API Key**: Pass `X-API-Key` header on every `/api/v1/*` request.
    - **JWT Bearer**: After login, use `Authorization: Bearer <access_token>` for protected endpoints.
    - **Admin**: Pass `X-Admin-Secret` header for `/api/admin/*` management endpoints.
  version: "1.0.0"
  contact:
    name: Security Verification System
servers:
  - url: http://localhost:5000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key (starts with sk_live_)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret for management endpoints
...
paths:
...
  /api/admin/apps/{app_id}/keys:
    post:
      tags: [Admin]
      summary: Generate API key for app (shown once)
      security:
        - AdminAuth: []
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
      responses:
        "201":
          description: Key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  key:
                    type: string
                    description: Full API key (shown only once)
                  prefix:
                    type: string
                  message:
                    type: string
  /api/admin/apps/{app_id}/keys/{prefix}:
    delete:
      tags: [Admin]
      summary: Revoke an API key
      security:
        - AdminAuth: []
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: integer
        - name: prefix
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
        "404":
          description: Key not found
...